refactor: use same control cloud for log pages
