+++ apps/dashboard/app/(app)/logs/components/table/hooks/use-logs-query.ts
@@ -2,11 +2,11 @@ import { trpc } from "@/lib/trpc/client";
 import type { Log } from "@unkey/clickhouse/src/logs";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import type { z } from "zod";
+import { HISTORICAL_DATA_WINDOW } from "../../../constants";
 import { useFilters } from "../../../hooks/use-filters";
 import type { queryLogsPayload } from "../query-logs.schema";
 
 // Duration in milliseconds for historical data fetch window (12 hours)
-export const HISTORICAL_DATA_WINDOW = 12 * 60 * 60 * 1000;
 type UseLogsQueryParams = {
   limit?: number;
   pollIntervalMs?: number;
