+++ apps/dashboard/app/(app)/logs/constants.ts
@@ -7,3 +7,5 @@ export const RED_STATES = ["DISABLED", "FORBIDDEN", "INSUFFICIENT_PERMISSIONS"];
 
 export const METHODS = ["GET", "POST", "PUT", "DELETE", "PATCH"] as const;
 export const STATUSES = [200, 400, 500] as const;
+
+export const HISTORICAL_DATA_WINDOW = 12 * 60 * 60 * 1000;
