+++ apps/dashboard/app/(app)/ratelimits/[namespaceId]/logs/components/table/hooks/use-logs-query.ts
@@ -1,11 +1,11 @@
 import { trpc } from "@/lib/trpc/client";
 import type { RatelimitLog } from "@unkey/clickhouse/src/ratelimits";
 import { useCallback, useEffect, useMemo, useState } from "react";
+import { HISTORICAL_DATA_WINDOW } from "../../../constants";
 import { useFilters } from "../../../hooks/use-filters";
 import type { RatelimitQueryLogsPayload } from "../query-logs.schema";
 
 // Duration in milliseconds for historical data fetch window (12 hours)
-export const HISTORICAL_DATA_WINDOW = 12 * 60 * 60 * 1000;
 type UseLogsQueryParams = {
   limit?: number;
   pollIntervalMs?: number;
