+++ apps/dashboard/app/(app)/ratelimits/[namespaceId]/logs/constants.ts
@@ -3,3 +3,5 @@ export const DEFAULT_STATUS_FLAG = 0;
 export const DEFAULT_DRAGGABLE_WIDTH = 500;
 export const MAX_DRAGGABLE_WIDTH = 800;
 export const MIN_DRAGGABLE_WIDTH = 300;
+
+export const HISTORICAL_DATA_WINDOW = 12 * 60 * 60 * 1000;
